+++ src/stores/settings.spec.ts
@@ -13,7 +13,9 @@ describe('settings store', () => {
   describe('fetch settings', () => {
     const settings: ApplicationSettings = {
       camera: {
+        pictureQuality: 60,
         shotTypes: ['pictures', 'videos'],
+        videoQuality: 90,
       },
       general: {
         deviceName: 'n',
@@ -51,14 +53,18 @@ describe('settings store', () => {
   describe('update persistent settings', () => {
     it("changes persistent settings' value", () => {
       const deviceName = 'd'
+      const pictureQuality = 40
       const sensitivity = 1
       const shotTypes: ShotType[] = ['pictures', 'videos']
       const siteName = 's'
       const timeZone = 't'
+      const videoQuality = 90
       const store = useSettingsStore()
       store.updatePersistentSettings({
         camera: {
           shotTypes,
+          pictureQuality,
+          videoQuality,
         },
         general: {
           deviceName,
@@ -69,7 +75,9 @@ describe('settings store', () => {
           sensitivity,
         },
       })
+      expect(store.camera.pictureQuality).toStrictEqual(pictureQuality)
       expect(store.camera.shotTypes).toStrictEqual(shotTypes)
+      expect(store.camera.videoQuality).toStrictEqual(videoQuality)
       expect(store.general.deviceName).toBe(deviceName)
       expect(store.general.siteName).toBe(siteName)
       expect(store.general.timeZone).toBe(timeZone)
@@ -77,19 +85,23 @@ describe('settings store', () => {
     })
   })
 
-  describe('upload persistent settings', () => {
+  describe('when updating only part of the settings', () => {
     const patchSettingsSpy = jest
       .spyOn(ApiClientService, 'patchSettings')
       .mockResolvedValue()
 
     it('uploads all persistent settings', async () => {
       const deviceName = 'd'
+      const pictureQuality = 40
       const sensitivity = 1
       const shotTypes: ShotType[] = ['pictures', 'videos']
       const siteName = 's'
       const store = useSettingsStore()
       const timeZone = 't'
+      const videoQuality = 90
+      store.camera.pictureQuality = pictureQuality
       store.camera.shotTypes = shotTypes
+      store.camera.videoQuality = videoQuality
       store.general.deviceName = deviceName
       store.general.siteName = siteName
       store.general.timeZone = timeZone
@@ -97,13 +109,66 @@ describe('settings store', () => {
       await store.uploadPersistentSettings()
       expect(ApiClientService.patchSettings).toHaveBeenCalledWith({
         camera: {
+          pictureQuality,
           shotTypes,
+          videoQuality,
         },
         general: {
           deviceName,
           siteName,
           timeZone,
         },
+        triggering: {
+          sensitivity,
+        },
+      })
+    })
+
+    it('uploads all camera settings', async () => {
+      const pictureQuality = 40
+      const shotTypes: ShotType[] = ['pictures', 'videos']
+      const videoQuality = 90
+      const store = useSettingsStore()
+      store.camera.pictureQuality = pictureQuality
+      store.camera.shotTypes = shotTypes
+      store.camera.videoQuality = videoQuality
+      await store.uploadAllCameraSettings()
+      expect(ApiClientService.patchSettings).toHaveBeenCalledWith({
+        camera: {
+          pictureQuality,
+          shotTypes,
+          videoQuality,
+        },
+      })
+    })
+
+    it('uploads all general settings', async () => {
+      const deviceName = 'd'
+      const siteName = 's'
+      const store = useSettingsStore()
+      const systemTime = new Date().toString()
+      const timeZone = 't'
+      store.general.deviceName = deviceName
+      store.general.siteName = siteName
+      store.general.systemTime = systemTime
+      store.general.timeZone = timeZone
+      await store.uploadAllGeneralSettings()
+      expect(ApiClientService.patchSettings).toHaveBeenCalledWith({
+        general: {
+          deviceName,
+          siteName,
+          systemTime,
+          timeZone,
+        },
+      })
+    })
+
+    it('uploads all trigger settings', async () => {
+      const sensitivity = 1
+      const store = useSettingsStore()
+      store.triggering.sensitivity = sensitivity
+      await store.uploadAllTriggerSettings()
+      expect(ApiClientService.patchSettings).toHaveBeenCalledWith({
         triggering: {
           sensitivity,
         },
@@ -121,11 +186,15 @@ describe('settings store', () => {
       .mockResolvedValue()
 
     it('uploads all settings', async () => {
+      const pictureQuality = 40
       const shotTypes: ShotType[] = ['pictures', 'videos']
       const systemTime = new Date().toString()
+      const videoQuality = 90
       const settings = {
         camera: {
+          pictureQuality,
           shotTypes,
+          videoQuality,
         },
         general: {
           deviceName: 'd',
@@ -138,7 +207,9 @@ describe('settings store', () => {
         },
       }
       const store = useSettingsStore()
+      store.camera.pictureQuality = pictureQuality
       store.camera.shotTypes = shotTypes
+      store.camera.videoQuality = videoQuality
       store.general.deviceName = settings.general.deviceName
       store.general.siteName = settings.general.siteName
       store.general.systemTime = systemTime