+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Introduce setting `trigger sensitivity`
 - Show dialog when time of device is more than 1 min off browser time
 - Send dedicated delete all files request when all files are selected
+- Introduce settings `picture quality` and `movie quality`
 
 ### Changed